feature(ui): fix header and sidebar
- make header fixed
- make sidebar fixed
- create subNav component
- hide subNav when scrolling down
- show subNav when scrolling up

[finishes #165852928]
