+++ client/assets/pages/_events-page.scss
@@ -21,6 +21,14 @@
     grid-column: sidebar-start/sidebar-end;
     width: rem(330px);
     vertical-align: middle;
+    &-fixed {
+      position: fixed;
+      width: 20.625rem;
+      top: 5.2rem;
+      @include mobile(600px) {
+        position: initial;
+      }
+    }
   }
 
   &__gallery {
@@ -204,8 +212,7 @@ main {
   width: 100%;
   background: var(--neutral-color);
   border-radius: 10px;
-  height: rem(373px);
-
+  min-height: rem(300px);
   .header {
     text-transform: uppercase;
     font-size: 15px;
@@ -331,7 +338,7 @@ main {
     width: 100%;
     background: var(--neutral-color);
     border-radius: 10px;
-    margin-bottom: 40px;
+    margin-bottom: 18px;
     padding: 1rem;
   }
   &__title {
